Extract the version checking logic to the dedicated layer
